+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/xpath/internal/ExpressionPrinter.java
@@ -13,8 +13,14 @@ import net.sf.saxon.om.Axis;
 
 /**
  * Simple printer for saxon expressions. Might be useful for debugging / during development.
+ *
+ * <p>Example:
+ * <pre>
+ * ExpressionPrinter printer = new ExpressionPrinter();
+ * printer.visit(query.xpathExpression.getInternalExpression());
+ * </pre>
  */
-public class ExpressionPrinter extends Visitor {
+public class ExpressionPrinter extends SaxonExprVisitor {
     private int depth = 0;
 
     private void print(String s) {