+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/table/internal/SymbolTableResolver.java
@@ -5,13 +5,9 @@
 
 package net.sourceforge.pmd.lang.java.symbols.table.internal;
 
-import static net.sourceforge.pmd.lang.java.symbols.table.internal.TypeOnlySymTable.localClassesOf;
 import static net.sourceforge.pmd.lang.java.symbols.table.internal.TypeOnlySymTable.nestedClassesOf;
 import static net.sourceforge.pmd.lang.java.symbols.table.internal.VarOnlySymTable.formalsOf;
-import static net.sourceforge.pmd.lang.java.symbols.table.internal.VarOnlySymTable.resourceIds;
-import static net.sourceforge.pmd.lang.java.symbols.table.internal.VarOnlySymTable.varsOfBlock;
 import static net.sourceforge.pmd.lang.java.symbols.table.internal.VarOnlySymTable.varsOfInit;
-import static net.sourceforge.pmd.lang.java.symbols.table.internal.VarOnlySymTable.varsOfSwitchBlock;
 
 import java.util.List;
 import java.util.Map;
@@ -27,11 +23,16 @@ import net.sourceforge.pmd.lang.java.ast.ASTForStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTForeachStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTImportDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTLambdaExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTLocalClassStatement;
+import net.sourceforge.pmd.lang.java.ast.ASTLocalVariableDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTModifierList;
 import net.sourceforge.pmd.lang.java.ast.ASTRecordConstructorDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTResource;
 import net.sourceforge.pmd.lang.java.ast.ASTResourceList;
+import net.sourceforge.pmd.lang.java.ast.ASTStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTSwitchExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTSwitchFallthroughBranch;
 import net.sourceforge.pmd.lang.java.ast.ASTSwitchLike;
 import net.sourceforge.pmd.lang.java.ast.ASTSwitchStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTTryStatement;
@@ -39,7 +40,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTTypeBody;
 import net.sourceforge.pmd.lang.java.ast.InternalApiBridge;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.ast.SideEffectingVisitorAdapter;
-import net.sourceforge.pmd.lang.java.symbols.SymbolResolver;
+import net.sourceforge.pmd.lang.java.internal.JavaAstProcessor;
 import net.sourceforge.pmd.lang.java.symbols.table.JSymbolTable;
 
 
@@ -51,89 +52,22 @@ import net.sourceforge.pmd.lang.java.symbols.table.JSymbolTable;
  */
 public final class SymbolTableResolver {
 
-    private final SymbolTableHelper myResolveHelper;
-    private final ASTCompilationUnit root;
-    // current top of the stack
-    private JSymbolTable myStackTop;
-
-
-    /**
-     * Initialize this resolver using an analysis context and a root node.
-     *
-     * @param symResolver Resolver for external references
-     * @param root        The root node
-     */
-    public SymbolTableResolver(SymbolResolver symResolver,
-                               SemanticChecksLogger logger,
-                               ASTCompilationUnit root) {
-        this.root = root;
-        myResolveHelper = new SymbolTableHelper(root.getPackageName(), symResolver, logger);
-        // this is the only place pushOnStack can be circumvented
-        myStackTop = EmptySymbolTable.getInstance();
-
+    private SymbolTableResolver() {
+        // façade
     }
 
-    /**
-     * Start the analysis.
-     */
-    public void traverse() {
-        assert myStackTop instanceof EmptySymbolTable
-            : "Top should be an empty symtable when starting the traversal";
-
-        root.jjtAccept(new MyVisitor(), null);
-
-        assert myStackTop instanceof EmptySymbolTable
-            : "Unbalanced stack push/pop! Top is " + myStackTop;
+    public static void traverse(JavaAstProcessor processor, ASTCompilationUnit root) {
+        SymbolTableHelper helper = new SymbolTableHelper(root.getPackageName(), processor);
+        new MyVisitor(root, helper).traverse();
     }
 
-
-    /**
-     * Create a new symbol table using {@link TableLinker#createAndLink(JSymbolTable, SymbolTableHelper, Object)},
-     * linking it to the top of the stack as its parent.
-     *
-     * Pushes must naturally be balanced with {@link #popStack()} calls.
-     *
-     * @param data Additional param passed to the linker. Passing parameters
-     *             like this avoids having to create a capturing lambda.
-     *
-     * @return 1 if the table was pushed, 0 if not
-     */
-    private <T> int pushOnStack(TableLinker<T> tableLinker, T data) {
-        AbstractSymbolTable created = tableLinker.createAndLink(peekStack(), myResolveHelper, data);
-        return pushOnStack(created) ? 1 : 0;
-    }
-
-    private boolean pushOnStack(AbstractSymbolTable table) {
-        assert table.getParent() == peekStack() : "Wrong parent";
-        if (table.isPrunable()) {
-            return false; // and don't set the stack top
-        }
-        this.myStackTop = table;
-        return true;
-    }
-
-    private JSymbolTable popStack() {
-        JSymbolTable curTop = this.myStackTop;
-        this.myStackTop = curTop.getParent();
-        return curTop;
-    }
-
-    private void popStack(int times) {
-        while (times-- > 0) {
-            popStack();
-        }
-    }
-
-    private JSymbolTable peekStack() {
-        return this.myStackTop;
-    }
 
     @FunctionalInterface
     private interface TableLinker<T> {
 
         /**
          * Create a symbol table, given its parent, a helper instance,
-         * and some other data passed by {@link #pushOnStack(TableLinker, Object)}.
+         * and some other data passed by {@link MyVisitor#pushOnStack(TableLinker, Object)}.
          *
          * @param stackTop Top of the stack, becomes the parent of the new node
          * @param helper   Shared helper
@@ -142,12 +76,32 @@ public final class SymbolTableResolver {
         AbstractSymbolTable createAndLink(JSymbolTable stackTop, SymbolTableHelper helper, T data);
     }
 
-    private class MyVisitor extends SideEffectingVisitorAdapter<Void> {
+    private static class MyVisitor extends SideEffectingVisitorAdapter<Void> {
+
+        private final ASTCompilationUnit root;
+        private final SymbolTableHelper myResolveHelper;
+        private JSymbolTable myStackTop;
+
+        MyVisitor(ASTCompilationUnit root, SymbolTableHelper helper) {
+            this.root = root;
+            myResolveHelper = helper;
+            // this is the only place pushOnStack can be circumvented
+            myStackTop = EmptySymbolTable.getInstance();
+        }
 
 
-        // The AstAnalysisConfiguration is only used in the constructor
-        // The parameter on the visit methods is unnecessary
-        // TODO introduce another visitor with `void visit(Node);` signature
+        /**
+         * Start the analysis.
+         */
+        public void traverse() {
+            assert myStackTop instanceof EmptySymbolTable
+                : "Top should be an empty symtable when starting the traversal";
+
+            root.jjtAccept(this, null);
+
+            assert myStackTop instanceof EmptySymbolTable
+                : "Unbalanced stack push/pop! Top is " + myStackTop;
+        }
 
         @Override
         public void visit(ASTModifierList node, Void data) {
@@ -234,25 +188,7 @@ public final class SymbolTableResolver {
 
         @Override
         public void visit(ASTBlock node, Void data) {
-            int pushed = 0;
-            pushed += pushOnStack(VarOnlySymTable::new, varsOfBlock(node));
-            pushed += pushOnStack(TypeOnlySymTable::new, localClassesOf(node));
-            setTopSymbolTableAndRecurse(node);
-            popStack(pushed);
-        }
-
-        @Override
-        public void visit(ASTForeachStatement node, Void data) {
-            int pushed = pushOnStack(VarOnlySymTable::new, node.getVarId());
-            setTopSymbolTableAndRecurse(node);
-            popStack(pushed);
-        }
-
-        @Override
-        public void visit(ASTForStatement node, Void data) {
-            int pushed = pushOnStack(VarOnlySymTable::new, varsOfInit(node));
-            setTopSymbolTableAndRecurse(node);
-            popStack(pushed);
+            visitBlockLike(node);
         }
 
         @Override
@@ -268,14 +204,46 @@ public final class SymbolTableResolver {
         private void visitSwitch(ASTSwitchLike node) {
             setTopSymbolTable(node);
             visitSubtree(node.getTestedExpression());
+            visitBlockLike(stmtsOfSwitchBlock(node));
+        }
 
-            // since there's no node for the block we have to set children individually
-            int pushed = pushOnStack(VarOnlySymTable::new, varsOfSwitchBlock(node));
 
-            for (JavaNode notExpr : node.children().drop(1)) {
-                setTopSymbolTableAndRecurse(notExpr);
+        private void visitBlockLike(Iterable<? extends ASTStatement> node) {
+            /*
+             * Process the statements of a block in a sequence. Each local
+             * var/class declaration is only in scope for the following
+             * statements (and its own initializer).
+             */
+            int pushed = 0;
+            for (ASTStatement st : node) {
+                if (st instanceof ASTLocalVariableDeclaration) {
+                    pushed += pushOnStack(VarOnlySymTable::new, ((ASTLocalVariableDeclaration) st).getVarIds());
+                } else if (st instanceof ASTLocalClassStatement) {
+                    pushed += pushOnStack(TypeOnlySymTable::new, ((ASTLocalClassStatement) st).getDeclaration());
+                }
+
+
+                setTopSymbolTable(st);
+                st.jjtAccept(this, null);
             }
 
+            popStack(pushed);
+        }
+
+        @Override
+        public void visit(ASTForeachStatement node, Void data) {
+            // the varId is only in scope in the body and not the iterable expr
+            setTopSymbolTableAndRecurse(node.getIterableExpr());
+
+            int pushed = pushOnStack(VarOnlySymTable::new, node.getVarId());
+            node.getBody().jjtAccept(this, data);
+            popStack(pushed);
+        }
+
+        @Override
+        public void visit(ASTForStatement node, Void data) {
+            int pushed = pushOnStack(VarOnlySymTable::new, varsOfInit(node));
+            setTopSymbolTableAndRecurse(node);
             popStack(pushed);
         }
 
@@ -284,16 +252,16 @@ public final class SymbolTableResolver {
 
             ASTResourceList resources = node.getResources();
             if (resources != null) {
-                int pushed = pushOnStack(VarOnlySymTable::new, resourceIds(resources));
+                NodeStream<ASTStatement> union =
+                    NodeStream.union(
+                        stmtsOfResources(resources),
+                        // use the body instead of unwrapping it so
+                        // that it has the correct symbol table too
+                        NodeStream.of(node.getBody())
+                    );
+                visitBlockLike(union);
 
-                setTopSymbolTableAndRecurse(resources);
-
-                ASTBlock body = node.getBody();
-                body.jjtAccept(this, data);
-
-                popStack(pushed); // pop resources table before visiting catch & finally
-
-                for (Node child : body.asStream().followingSiblings()) {
+                for (Node child : node.getBody().asStream().followingSiblings()) {
                     ((JavaNode) child).jjtAccept(this, data);
                 }
             } else {
@@ -307,6 +275,9 @@ public final class SymbolTableResolver {
             setTopSymbolTableAndRecurse(node);
             popStack(pushed);
         }
+
+
+        // <editor-fold defaultstate="collapsed" desc="Stack manipulation routines">
 
         private void setTopSymbolTable(JavaNode node) {
             InternalApiBridge.setSymbolTable(node, peekStack());
@@ -322,6 +293,64 @@ public final class SymbolTableResolver {
                 child.jjtAccept(this, null);
             }
         }
+
+
+
+        /**
+         * Create a new symbol table using {@link TableLinker#createAndLink(JSymbolTable, SymbolTableHelper, Object)},
+         * linking it to the top of the stack as its parent.
+         *
+         * Pushes must naturally be balanced with {@link #popStack()} calls.
+         *
+         * @param data Additional param passed to the linker. Passing parameters
+         *             like this avoids having to create a capturing lambda.
+         *
+         * @return 1 if the table was pushed, 0 if not
+         */
+        private <T> int pushOnStack(TableLinker<T> tableLinker, T data) {
+            AbstractSymbolTable created = tableLinker.createAndLink(peekStack(), myResolveHelper, data);
+            return pushOnStack(created) ? 1 : 0;
+        }
+
+        private boolean pushOnStack(AbstractSymbolTable table) {
+            assert table.getParent() == peekStack() : "Wrong parent";
+            if (table.isPrunable()) {
+                return false; // and don't set the stack top
+            }
+            this.myStackTop = table;
+            return true;
+        }
+
+        private JSymbolTable popStack() {
+            JSymbolTable curTop = this.myStackTop;
+            this.myStackTop = curTop.getParent();
+            return curTop;
+        }
+
+        private void popStack(int times) {
+            while (times-- > 0) {
+                popStack();
+            }
+        }
+
+        private JSymbolTable peekStack() {
+            return this.myStackTop;
+        }
+
+        static NodeStream<ASTStatement> stmtsOfSwitchBlock(ASTSwitchLike node) {
+            return node.getBranches()
+                       .filterIs(ASTSwitchFallthroughBranch.class)
+                       .flatMap(ASTSwitchFallthroughBranch::getStatements);
+        }
+
+
+        static NodeStream<ASTLocalVariableDeclaration> stmtsOfResources(ASTResourceList node) {
+            return node.toStream().map(ASTResource::asLocalVariableDeclaration);
+        }
+
+
+        // </editor-fold>
+
     }
 
 }