+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/table/internal/VarOnlySymTable.java
@@ -15,7 +15,6 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import net.sourceforge.pmd.lang.ast.NodeStream;
-import net.sourceforge.pmd.lang.java.ast.ASTBlock;
 import net.sourceforge.pmd.lang.java.ast.ASTForStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTFormalParameter;
 import net.sourceforge.pmd.lang.java.ast.ASTLambdaExpression;
@@ -24,12 +23,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTLocalVariableDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTRecordComponent;
 import net.sourceforge.pmd.lang.java.ast.ASTRecordConstructorDeclaration;
-import net.sourceforge.pmd.lang.java.ast.ASTResource;
-import net.sourceforge.pmd.lang.java.ast.ASTResourceList;
-import net.sourceforge.pmd.lang.java.ast.ASTSwitchFallthroughBranch;
-import net.sourceforge.pmd.lang.java.ast.ASTSwitchLike;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
-import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.symbols.JConstructorSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JFormalParamSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JVariableSymbol;
@@ -104,24 +98,6 @@ final class VarOnlySymTable extends AbstractSymbolTable {
         return new VarResolveResult(it.getSymbol(), this, it);
     }
 
-    @NonNull
-    protected ResolveResultImpl<JVariableSymbol> makeResult(@NonNull JVariableSymbol it, JavaNode contributor) {
-        return new VarResolveResult(it, this, contributor);
-    }
-
-    static NodeStream<ASTVariableDeclaratorId> varsOfBlock(ASTBlock node) {
-        return node.children(ASTLocalVariableDeclaration.class)
-                   .flatMap(ASTLocalVariableDeclaration::getVarIds);
-    }
-
-    static NodeStream<ASTVariableDeclaratorId> varsOfSwitchBlock(ASTSwitchLike node) {
-        return node.getBranches()
-                   .filterIs(ASTSwitchFallthroughBranch.class)
-                   .flatMap(ASTSwitchFallthroughBranch::getStatements)
-                   .filterIs(ASTLocalVariableDeclaration.class)
-                   .flatMap(ASTLocalVariableDeclaration::getVarIds);
-    }
-
     static NodeStream<ASTVariableDeclaratorId> varsOfInit(ASTForStatement node) {
         return NodeStream.of(node.getInit())
                          .filterIs(ASTLocalVariableDeclaration.class)
@@ -141,11 +117,4 @@ final class VarOnlySymTable extends AbstractSymbolTable {
     static NodeStream<ASTVariableDeclaratorId> formalsOf(ASTMethodOrConstructorDeclaration node) {
         return node.getFormalParameters().toStream().map(ASTFormalParameter::getVarId);
     }
-
-    static NodeStream<ASTVariableDeclaratorId> resourceIds(ASTResourceList node) {
-        return node.toStream()
-                   .map(ASTResource::asLocalVariableDeclaration)
-                   .flatMap(ASTLocalVariableDeclaration::getVarIds);
-    }
-
 }